chore(evm): cleanup EVM files
